Refactor upload file code for chatroom and Upload screen
