+++ screens/ChatRoom/index.tsx
@@ -111,6 +111,15 @@ interface ChatRoom {
   route: any;
 }
 
+interface UploadResource {
+  selectedImages: any;
+  conversationID: any;
+  chatroomID: any;
+  selectedFilesToUpload: any;
+  uploadingFilesMessages: any;
+  isRetry: boolean;
+}
+
 const ChatRoom = ({navigation, route}: ChatRoom) => {
   const flatlistRef = useRef<any>(null);
   let refInput = useRef<any>();
@@ -1385,10 +1394,21 @@ const ChatRoom = ({navigation, route}: ChatRoom) => {
     setDMBlockAlertModalVisible(false);
   };
 
-  const uploadResource = async (selectedImages: any, conversationID: any) => {
+  const uploadResource = async ({
+    selectedImages,
+    conversationID,
+    chatroomID,
+    selectedFilesToUpload,
+    uploadingFilesMessages,
+    isRetry,
+  }: UploadResource) => {
     for (let i = 0; i < selectedImages?.length; i++) {
-      let attachmentType = selectedImages[i]?.type;
-      let docAttachmentType = selectedImages[i]?.type;
+      let attachmentType = isRetry
+        ? selectedImages[i]?.type
+        : selectedImages[i]?.type?.split('/')[0];
+      let docAttachmentType = isRetry
+        ? selectedImages[i]?.type
+        : selectedImages[i]?.type?.split('/')[1];
       let thumbnailURL = selectedImages[i]?.thumbnail_url;
       let name =
         attachmentType === IMAGE_TEXT
@@ -1423,7 +1443,7 @@ const ChatRoom = ({navigation, route}: ChatRoom) => {
         Key: thumbnailUrlPath,
         Body: thumbnailUrlImg,
         ACL: 'public-read-write',
-        ContentType: selectedImages[i]?.type, // Replace with the appropriate content type for your file
+        ContentType: 'image/jpeg', // Replace with the appropriate content type for your file
       };
 
       try {
@@ -1450,16 +1470,22 @@ const ChatRoom = ({navigation, route}: ChatRoom) => {
             conversation_id: conversationID,
             files_count: selectedImages?.length,
             index: i,
-            meta: {
-              size:
-                docAttachmentType === PDF_TEXT
-                  ? selectedImages[i]?.size
-                  : selectedImages[i]?.fileSize,
-            },
+            meta:
+              fileType === VIDEO_TEXT
+                ? {
+                    size: selectedFilesToUpload[i]?.fileSize,
+                    duration: selectedFilesToUpload[i]?.duration,
+                  }
+                : {
+                    size:
+                      docAttachmentType === PDF_TEXT
+                        ? selectedFilesToUpload[i]?.size
+                        : selectedFilesToUpload[i]?.fileSize,
+                  },
             name:
               docAttachmentType === PDF_TEXT
-                ? selectedImages[i]?.name
-                : selectedImages[i]?.fileName,
+                ? selectedFilesToUpload[i]?.name
+                : selectedFilesToUpload[i]?.fileName,
             type: fileType,
             url: awsResponse,
             thumbnail_url:
@@ -1479,6 +1505,7 @@ const ChatRoom = ({navigation, route}: ChatRoom) => {
             ID: conversationID,
           },
         });
+        return error;
       }
       dispatch({
         type: CLEAR_SELECTED_FILES_TO_UPLOAD,
@@ -1496,7 +1523,7 @@ const ChatRoom = ({navigation, route}: ChatRoom) => {
     });
   };
 
-  const handleFileUpload = async (conversationID: any) => {
+  const handleFileUpload = async (conversationID: any, isRetry: any) => {
     let selectedFilesToUpload = uploadingFilesMessages[conversationID];
     dispatch({
       type: SET_FILE_UPLOADING_MESSAGES,
@@ -1508,10 +1535,14 @@ const ChatRoom = ({navigation, route}: ChatRoom) => {
         ID: conversationID,
       },
     });
-    const res = await uploadResource(
-      selectedFilesToUpload?.attachments,
-      conversationID,
-    );
+    const res = await uploadResource({
+      selectedImages: selectedFilesToUpload?.attachments,
+      conversationID: conversationID,
+      chatroomID: chatroomID,
+      selectedFilesToUpload: selectedFilesToUpload,
+      uploadingFilesMessages,
+      isRetry: isRetry,
+    });
     return res;
   };
 
@@ -1665,6 +1696,7 @@ const ChatRoom = ({navigation, route}: ChatRoom) => {
                 navigation={navigation}
                 isUploadScreen={false}
                 myRef={refInput}
+                handleFileUpload={handleFileUpload}
               />
             ) : user.state !== 1 && chatroomDetails?.chatroom.type === 7 ? (
               <View style={styles.disabledInput}>
@@ -1803,6 +1835,7 @@ const ChatRoom = ({navigation, route}: ChatRoom) => {
               isUploadScreen={false}
               isPrivateMember={chatroomDetails?.chatroom?.is_private_member}
               myRef={refInput}
+              handleFileUpload={handleFileUpload}
             />
           ) : (
             <View style={styles.disabledInput}>
